Put metrics in module
