+++ policy-controller/k8s/index/src/outbound.rs
@@ -1,3 +1,3 @@
 pub mod index;
 
-pub use index::{Index, ServiceRef, SharedIndex};
+pub use index::{metrics, Index, ServiceRef, SharedIndex};
